[clear] user周りのAPIリファクタ
